requirements.txt
notebook
ipywidgets
ipykernel
tqdm
pandas
scikit-learn
openai
qdrant-client>=1.14.2
fastembed
python-dotenv
python-docx
requests
streamlit
sentence_transformers
ray
mlflow
duckdb